Cache refresh token to config file
