+++ cmd/requests_patch.go
@@ -5,9 +5,10 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"github.com/spf13/cobra"
 	"goft/pkg/ftapi"
 	"io/ioutil"
+
+	"github.com/spf13/cobra"
 )
 
 // NewRequestsPatchCmd Create the patch request cmd
@@ -15,15 +16,16 @@ func NewRequestsPatchCmd(api *ftapi.APIInterface) *cobra.Command {
 	return &cobra.Command{
 		Use:   "patch path",
 		Short: "Send a PATCH request",
-		Long: `Send a PATCH request to path, the path must be the part after /v2/`,
-		Args: cobra.ExactArgs(1),
+		Long:  `Send a PATCH request to path, the path must be the part after /v2/`,
+		Args:  cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
+			defer saveConfig()
 			fmt.Print("Enter payload: (press CTRL+D when you finish)\n")
 			var payload string
 			scanner := bufio.NewScanner(cmd.InOrStdin())
 			for scanner.Scan() {
 				line := scanner.Text()
-				payload += line+"\n"
+				payload += line + "\n"
 			}
 			fmt.Println(payload)
 			resp, err := (*api).Patch(args[0], "application/json", bytes.NewReader([]byte(payload)))
@@ -42,6 +44,7 @@ func NewRequestsPatchCmd(api *ftapi.APIInterface) *cobra.Command {
 		},
 	}
 }
+
 var requestsPatchCmd = NewRequestsPatchCmd(&API)
 
 func init() {
